Move streaming metrics to two nodes tests into their own classes. Make the StreamingMetricsTestUtils not inherit from TestBaseImpl
